backend/pyproject.toml
[project]
name = "taskly-backend"
version = "1.0.0"
description = "Taskly task management backend — FastAPI + SQLite"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "itsdangerous>=2.2.0",
    "python-multipart>=0.0.9",
    "bcrypt>=5.0.0",
    "libsql-client>=0.3.1",
    "mangum>=0.21.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
]
